+++ typescript/registry/RELEASES.md
@@ -107,3 +107,13 @@ Based on:
 - [typescript v0.6.1] typescript/registry
 ### Releases
 - [NPM v0.6.1] https://www.npmjs.com/package/@smithery/registry/v/0.6.1 - typescript/registry
+
+## 2025-11-30 00:11:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.650.0 (2.743.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.2] typescript/registry
+### Releases
+- [NPM v0.6.2] https://www.npmjs.com/package/@smithery/registry/v/0.6.2 - typescript/registry